requirements.txt
openai>=1.0.0
pyyaml>=6.0
requests>=2.28.0
